+++ feature/src/main/res/values/strings.xml
@@ -32,6 +32,7 @@
     <string name="source_tagline_outline">Your wiki</string>
     <string name="source_tagline_mempalace">Your palace</string>
     <string name="source_tagline_radio">Live radio</string>
+    <string name="source_tagline_techempower">Guides &amp; resources</string>
     <string name="source_tagline_notion">Your Notion</string>
     <string name="source_tagline_hackernews">Tech news</string>
     <string name="source_tagline_arxiv">Research papers</string>
